+++ example/a.ts
@@ -1,18 +1,9 @@
 import { Elysia, t } from '../src'
 
-const ws = new Elysia().ws('/plugin', {
-	message(ws, message) {
-		ws.send(message)
-	}
-})
+const plugin = async () => new Elysia().get('/', () => 'A')
 
 const app = new Elysia()
-	.use(ws)
-	.ws('/', {
-		message(ws, message) {
-			ws.send(message)
-		}
-	})
+	.use(plugin())
 	.listen(3000)
 
 console.log(app.routes)